Working on putting on tighter tracks
When a hinge joint is copied, it's stretched status is saved as the default one, allowing for further stretching when physics are applied to it. To counteract this I am trying to create a funtional idler that moves to proper position once the game loads and am disabling gravity on the track links and enabling it once the game loads.